rekenen-ai-prompts/requirements.txt
# Core dependencies
pyyaml>=6.0
pydantic>=2.0

# Testing
pytest>=7.0
pytest-cov>=4.0

# Utilities
python-dotenv>=1.0
